[AI-FSSDK] [FSSDK-12337] Fix test structure, mypy and ruff compliance
- Move feature rollout tests from standalone test_feature_rollout.py into
  test_config.py following module-level testing convention
- Use base.BaseTest instead of unittest.TestCase for consistency
- Fix mypy strict type errors in Variation construction using cast
- All checks pass: ruff, mypy --strict, pytest (941/941)
